requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
argon2-cffi
psycopg2-binary
pyJWT
gunicorn
python-dotenv

